+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog
 - Added list program account receipt endpoint
 - Added list user receipt endpoint
 - Added list prepaid card receipt endpoint
+- Fixed the payment representation (renamed description to notes)
 
 0.1.0 (2016-06-30)
 ------------------